+++ loaders/passport.js
@@ -1,46 +1,62 @@
 const passport = require("passport");
 const LocalStrategy = require("passport-local");
-const db = require("../db");
-const bcrypt = require("bcrypt");
-const { findOneByEmail } = require("../helperFunctions/index");
+const session = require("express-session");
+const pgSession = require("connect-pg-simple")(session);
+const pg = require("pg");
+const { findOneByEmail, compareIt } = require("../helperFunctions/index");
+
+//connect with PostgresSQL
+const connectionString = {
+  connectionString: process.env.DATABASE_URL,
+};
+
+const pgPool = new pg.Pool(connectionString);
 
 module.exports = (app) => {
+  // Creates a session
+  app.use(
+    session({
+      secret: process.env.SESSION_SECRET,
+      resave: true,
+      saveUninitialized: true,
+      cookie: {
+        secure: true,
+        maxAge: 24 * 60 * 60 * 1000,
+      },
+      store: new pgSession({
+        pool: pgPool, // Connection pool
+        tableName: "user_sessions", // Use another table-name than the default "session" one
+        // Insert connect-pg-simple options here
+      }),
+    })
+  );
   // Initialize passport
   app.use(passport.initialize());
   app.use(passport.session());
 
   // Set method to serialize data to store in cookie
-  passport.serializeUser((user, done) => {
-    // done(null, user.id);
-    done(null, { id: user.id, username: user.email });
+  passport.serializeUser(function (user, done) {
+    done(null, user.id);
   });
 
   // Set method to deserialize data stored in cookie and attach to req.user
-  passport.deserializeUser((id, done) => {
+  passport.deserializeUser(function (id, done) {
     done(null, { id });
   });
 
   // Configure strategy to be use for local login
   passport.use(
     new LocalStrategy(async (username, password, done) => {
-      //find user in database
-
-      const user = await findOneByEmail(username);
-
-      // Passwort prüfen
-      async function compareIt(password, userDb) {
-        const validPassword = await bcrypt.compare(password, userDb.password);
-        const passwordIsCorrect = await validPassword;
-        return passwordIsCorrect;
-      }
-
       try {
+        //find user in database
+        const user = await findOneByEmail(username);
+        if (!user) {
+          return done(null, false);
+        }
+        // Passwort prüfen
         const checkPassword = await compareIt(password, user);
-
         if (!checkPassword) {
-          return done(null, false, {
-            message: "Incorrect username or password.",
-          });
+          return done(null, false);
         }
         return done(null, user);
       } catch (err) {
